checking private repos moved to jobs.py
